+++ sync/SyncDataCollection.js
@@ -148,7 +148,7 @@ var SyncDataCollection = (function () {
         var localColl = syncSetting.localCollection;
         return this.syncUpAndUpdateCollection(localColl, primaryKey, primaryKeys, function convertAndSendItemsToServer(items) {
             var beforeSyncUpPrepTimer = self.notifyActionStart ? self.notifyActionStart("beforeSyncUpPrep", localColl) : null;
-            var toSvcObj = syncSetting.convertToSvcObjectFunc;
+            var toSvcObj = syncSetting.toSvcObject;
             var data = null;
             if (primaryKey) {
                 data = SyncDataCollection.checkAndConvertSingleKeyItems(localColl.getName(), items, primaryKeyChecker, toSvcObj);
@@ -261,7 +261,7 @@ var SyncDataCollection = (function () {
         return function addUpdateOrRemoveItemsFunc(items) {
             var table = syncSettings.localCollection;
             var findFilterFunc = syncSettings.findFilterFunc;
-            var convertToLocalObjectFunc = syncSettings.convertToLocalObjectFunc;
+            var convertToLocalObjectFunc = syncSettings.toLocalObject;
             if (syncDownOp.removeAll) {
                 table.clearCollection();
             }